emotelist server selection
emotelist.js command now lets the user select the server they wish to show the emotes for. They are only allowed to select from the servers they share with the bot, however.

Modified the display of the emotes to add clarity when showing emotes for multiple servers in a single channel.