typescript conversion, main files
index.ts and the event files have all been converted to TypeScript files. Transpile them into JavaScript files by running "npm run build" in root folder.

Typing the reaction roles, twitter uses, and bot reaction appear to be useful for clarity when editing data. Uncertain of the effect of TypeScript on runtimes, as this is an experimental branch.