Bug fix: roleupdate.ts
Fixed bug where role was not actually being updated in the server. Resolved by editing the guild role directly-output by a Promise, rather than assigning the guild role to a constant and editing that.

Fixed bug where the role is given ALL of the permissions when ADMINISTRATOR is removed, even if the role did not have it before. Fixed by checking the actual role permissions bitfield, not the effective permissions array.

Command now allows the user to modify the position and hoist properties of the role. The bot cannot move the role above itself (the max is stated)