gem command JS -> TS
Converted the last JS command, gem.ts, to TS.

The embed messages for each gem are now stored in a TS file, src/info/wiki/gems.ts, which is sent whenever an input matches the key of the collection.

This, however, means that aliases do not work with the command. To compensate, the name of all available gems are displayed when no argument is given.