Folder cleaning, attempted bugfix
Moved reception.ts to info/server folder to be align with other reference files.

Created a welcome.ts file to store the embed information for a welcome message. Includes log and welcome channels as outputs.

guildMemberAdd is not emitted when a person leaves and rejoins the server. Unsure if this is because they were in the server previously. Otherwise, the event functions OK when it does emit.