partial "emote subscriber" event
Deleted emitadd.ts, to be replaced by the events command emit.ts, which allows the user to emit a variety of client events in a messaged server.

Created a reference file, emotechannel.ts (maybe renamed later). It contains info of a server that seeks to "subscribe" to the emotes of other servers.

The emojiCreate.ts event is triggered when an emote is added to a server. It reads emotechannel.ts and checks if any servers are "subscribed" to it. If there are, it attempts to delete the messages in a channel and update all of the emote lists in that channel.

Also added a line to delete the triggering message of the emotelist.ts command.

emojiCreate.s