Additional filter in message.js emote replace
Moved the filter to retrieve the author's user guilds (out of the bot's cached guilds) to a standalone function, functions/filters.ts.

This is used in both emotelist.ts and message.ts. The functionality is the same in emotelist.ts.

In message.ts, the emote replace block still prioritizes the messaged guild. Next, however, it prioritizes the first match found in the user's guilds. If no matches were found, it then attempts to search all client guilds, as usual.

There is currently no method to "prioritize" which emotes to search from first, except by defining the order to search in by a separate filter function.