Refactoring source files
Functionality restored to the bot on all commands and events after changes to reference files.

Client now uses partials, which removes the need to cache old messages in ready.ts. As a result, the IDs of messages used for reaction roles no longer have to be manually-updated in the reference file. Instead, it relies on the checks for the existence of an embed message, an embed title matching one defined in the reference file, and for the message to exist in the channel designated by the reference file.

botStatuses.ts has been updated to show the various statuses more clearly. The client now randomly-chooses a status via a function.

Most of the guild definitions have been consolidated into the single reference file guilds.ts. Seeking to consolidate reception.ts into guilds.ts at a later date.