ready.ts & message.ts handlers
Methods in ready.ts and message.ts have been converted into individual handler files, stored in src/handlers, that are imported to their respective events.

Added method in parsers.ts to search for a emoji from the client's cache given a name, which is used in message.ts to search for emojis beginning from the messaged guild.

handleSpamEmojiChannels.ts and spamChannels.ts have been added to automatically remove & log a message if it does not contain the correct emojis in specific channels.

If a user uses the non-nitro emoji command (e.g. for animated emotes), they will NOT be warned if the output message is valid for the spamming channel, although the message will deleted as usual.