Initial upload for typeorm / mysql database
mysql database to be manipulated with typeorm. Reference files updated.

db.ts exports a function, prepareConnection(), that closes any existing database connections prior to connecting to a database for use in any other file.

ready.ts contains a possible block where client guilds are added to a table, if they do not already exist in the database.