handlers for emotelist, emojiCreate, guildMemberAdd
Handlers created for emotelist, emojiCreate, and guildMemberAdd. Handlers required for sending emote lists is now outside of the emotelist.ts command file.

In collectorFilters.ts, Message objects are used it allows CommandoMessage to also be inputs.

Handler for sending emote list now sends the emotes in large message groups containing up to 25 emotes, which significantly the time required to bulk delete and send the emote lists.