bugfix on message deletion
Before deleting a message (original or webhook), the client first checks if it is already deleted.

Additional if-block now deletes the nonNitroWebhookMessage if it does not meet the content criteria of a spam channel.

In general, messages are now deleted if the handleNonNitroEmoji handler runs successfully, provided they were not deleted already.