handling roledel, cleaning
Handlers, parsers, and filters updated to work for roledel.

roleadd.ts and roledel.ts now use a validator which checks the input name against the messaged guild. For roledel, the parser then returns the Role for later use.

Added an optional role parameter to handleRoleData so that the embed description contains the mentioned version of the role, as used in roledel.

Removed unused variables in DEFAULT_ROLE.ts