handling roleupdate, update other role commands
roleupdate.ts now uses a validator, parser, and handlers like the other role commands, with minor differences in implementation.

Getting role data from a Role object is exported to roledel.ts and roleupdate.ts

The permissions field in RoleData is typed as an array, which simplifies editing the contents when adding or updating a role.