type safety in handleRoleDataEdit
In handleRoleDataEdit.ts, used if-statements to check the existence of a property (i.e. not null or not undefined) in the args: RoleData input.

This removes the need to typecast the values in args, although the permissions must still be typecast from PermissionResolveable so that they can be modified like an array.

Updated import in handleRoleDataConfirmation.ts