Enums for argument validation
CLIENT_EVENTS.ts has been repurposed into a custom-types file. ClientEvents & RolePermissions are enums that can be used to validate inputs to emit.ts (for client events) and handleRoleDatargs (for editing role permissions).

The emit.ts command also displays the list of client events to input. The input is also validated against the ClientEvents enum.

roleadd.ts and roleupdate.ts also the same prompt for the role arguments which displays list of available permissions to modify. Default permissions are not specified.

Note: not all events are used in the command, so the table may be shortened at a later date.