Initial upload for database
mysql database appears to function alongside typeorm. Initial upload to provide a checkpoint. Commit does not modify existing infrastructure for other commands.

db.ts remade as a handler that returns the database connection.

Created entities required for designating a reaction role within a server.

Various handlers made to request a single guild, reaction role, and reaction. These handlers also add the respective query to the database if they do not already exist.

reactionrole.ts command currently checks for a role in a guild and if an emoji has been provided to use as a reaction.

Currently unable to update the relations between the reaction role and reaction (emoji).