Entities update + localhost fixes
mysql server 8.0 / bin was unknown to command line because it was not added to system env variables PATH. terminal, including in VSCode, now recognizes the command, and the local database + columns can now be viewed.

Entities for the database have been updated to extend a Base entity, where applicable.

Because TYPEORM_SYNCHRONIZE = true in ormconfig.env, the database will be updated with any changes in the entities folder, which makes the info unsafe during development. For example, if an entity was modified (added, deleted, or has a changed column), the database will sync accordingly.

Consider using an dev database that copies the current state of the true database and only alters this.