Refactor reaction collectors
Note: This commit should be under the refactor branch or similar, but these changes resulted from using a reaction collector for a database command.

Many commands use a reaction collector, but the only differences are the collector.on callbacks and the reaction options. To reduce repeating code, these commands will now use a utility function to generate a collector (based on the input options) and then attach the callbacks to it (based on the callbacks required by the command).