Register database commands
Added a group for the database commands. Added a command that adds all client guilds to the database, if they do not exist. Should not overwrite existing guild data, due to the primary columns needing to be unique, but little testing outside of spamming the command has been performed.

All entities are now exported from an index.ts file inside of src\entity to clean up imports in other files.