database commands
types, interfaces, and constants updated to reflect re-used properties for the database.

handlers added to query all rows from a specified entity table and then display the values in an embed. commands queryAll and registerGuilds both use this embed.

reactionrole command now properly adds and deletes a role from the database. Currently, it does not