register objects into database, entity changes
Base entity changed to use a primary column over generated, as the id's for this class was always be a pre-determined id from the Discord objects.

As Embed entities must only have unique pairs of titles & channels, a primary generated column with uuid is used, which allows duplicate titles and channel ids (across guilds), provided the pairs are unique.

handleIDQuery command now types the return to improve the function info. Result must still be typecast based on the expected types + null + undefined.

registerGuilds is now an owner only command that force adds all guilds to the database.

registerGuild, registerChannel, and registerEmbed added as general commands (will later be restricted to users with MANAGE_CHANNELS permissions). These commands simply check if an existing entity exists with the input values, aborting if it does exist and inserts into database if not.

No command methods for removing entities from the database have been implemented at this time.

register commands required for the reaction role command (which will be renamed to be similar to these), as the reaction role and children will ultimately reference these entities.