registerReaction, entity changes
Embed messages and reactions now use a uuid as the primary generated column.

Embed messages do not have a unique Discord id to link to. Thus, its title and channel location are used to make it a unique entry.

Reactions may also be used in multiple cases, thus the embed message title and channel location are used to make it a unique entry.

registerReaction.ts now handles registering an emoji into the database to be used for reaction roles.