viewGuild command, entity changes
Updated channels column "embed" to "embeds", as it is supposed to be an array of embed entities.

register__ commands now reference the client directly from the command (this.client) instead of instanciating the CommandoMessage when calling the validate or parse functions.

Added a getGuild.ts util to get a guild from client based on an id or name input.

viewGuild.ts command allows the user to view the database entries for the guild. The owner may also specify any guild from the client.