Entities update
All entities now have a @ManyToOne relationship with the guild entry, as all of these entities are tied to a singular guild.

As reactions can only be used in reaction roles within their guild, reactions now use their Discord ID as the primary column.

In event a guild wants to reuse a reaction in multiple embed messages for roles, Embed and Reaction now share a @ManyToMany relationship. The Embed messages still must be distinct, however, based on the title & channel.