ConnectionClient
Created a new class extending the CommandoClient, which adds a database connection to the client. Also created a new ConnectionCommand that extends the Command. These allow a database connection to be passed around the events and commands, no longer requiring connecting & disconnecting to a database repeatedly.

Updates to other commands or events requiring a connection to follow.