addRecord.ts milestone
addRecord.ts can now successfully add a channel, reaction, or role to a database for guild. viewGuild.ts can now show the records that are registered for a guild (does not show relationships).

A lot of other files were moved, renamed, and modified. Many old commands / functions may have been broken in the process. To be determined.